refactor: Rename seam/ to http/
"Seam" names the whole product, not a layer. This layer is the SDK/HTTP edge, so it takes the SDK package's name (@seamapi/http): http/client.ts constructs SeamHttp, http/request.ts makes the request. TESTING.md references updated.
