+++ src/lib/interact/interact-for-acs-user.ts
@@ -1,4 +1,4 @@
-import { getSeam } from '../seam/client.js'
+import { getSeam } from '../http/client.js'
 import { interactForAcsSystem } from './interact-for-acs-system.js'
 import { interactForResource } from './interact-for-resource.js'
