+++ src/lib/interact/interact-for-action-attempt-poll.ts
@@ -1,9 +1,9 @@
 import type { ActionAttemptsGetResponse } from '@seamapi/http/connect'
 
-import { getSeam } from '../seam/client.js'
+import { getSeam } from '../http/client.js'
 import { getOutput } from '../output/get-output.js'
-import { promptConfirm } from './prompt.js'
 import { withLoading } from '../output/with-loading.js'
+import { promptConfirm } from './prompt.js'
 
 export const interactForActionAttemptPoll = async (
   actionAttempt: ActionAttemptsGetResponse['action_attempt'],
